+++ client/Packages/com.beamable.server/Editor/UI/Model/MicroserviceModel.cs
@@ -53,7 +53,7 @@ namespace Beamable.Editor.UI.Model
 
 		[field: SerializeField]
 		public ServiceStatus RemoteStatus { get; protected set; }
-
+		public override bool IsArchived => Config.Archived;
 		public MicroserviceConfigurationEntry Config => MicroserviceConfiguration.Instance.GetEntry(Descriptor.Name);
 		public List<MongoStorageModel> Dependencies { get; private set; } = new List<MongoStorageModel>(); // TODO: This is whacky.
 		public override bool IsRunning => ServiceBuilder?.IsRunning ?? false;
@@ -122,6 +122,7 @@ namespace Beamable.Editor.UI.Model
 			OnBuild?.Invoke(task);
 			return task;
 		}
+
 		public void OpenLocalDocs()
 		{
 			var de = BeamEditorContext.Default;
@@ -210,6 +211,16 @@ namespace Beamable.Editor.UI.Model
 			if (!AreLogsAttached)
 			{
 				evt.menu.BeamableAppendAction($"Reattach Logs", pos => AttachLogs());
+			}
+			
+			evt.menu.AppendSeparator();
+			if (Config.Archived)
+			{
+				evt.menu.AppendAction("Unarchive", _ => Unarchive());
+			}
+			else
+			{
+				evt.menu.AppendAction("Archive", _ => Archive());
 			}
 		}
 
@@ -326,6 +337,19 @@ $@"{{
 			}
 #endif
 		}
+
+		public void Archive()
+		{
+			Config.Archived = true;
+			MicroservicesDataModel.Instance.OnServiceArchived?.Invoke(this);
+		}
+
+		public void Unarchive()
+		{
+			Config.Archived = false;
+			MicroservicesDataModel.Instance.OnServiceUnarchived?.Invoke(this);
+		}
+		
 		public override void Refresh(IDescriptor descriptor)
 		{
 			// reset the descriptor and statemachines; because they aren't system.serializable durable.