+++ client/Packages/com.beamable.server/Editor/UI/Model/MicroservicesDataModel.cs
@@ -64,6 +64,9 @@ namespace Beamable.Editor.UI.Model
 		public Action<ServiceManifest> OnServerManifestUpdated;
 		public Action<GetStatusResponse> OnStatusUpdated;
 		public Guid InstanceId = Guid.NewGuid();
+		
+		public Action<ServiceModelBase> OnServiceArchived;
+		public Action<ServiceModelBase> OnServiceUnarchived;
 
 		public void RefreshLocal()
 		{
@@ -237,6 +240,9 @@ namespace Beamable.Editor.UI.Model
 		public bool ContainsRemoteOnlyModel(string serviceName) => AllRemoteOnlyServices?.Any(s => s.Descriptor.Name.Equals(serviceName)) ?? false;
 		public bool ContainsModel(string serviceName) => AllLocalServices?.Any(s => s.Descriptor.Name.Equals(serviceName)) ?? false;
 
+		public bool IsArchived(string serviceName) =>
+			AllLocalServices.First(s => s.Descriptor.Name.Equals(serviceName)).IsArchived;
+
 		public T GetModel<T>(IDescriptor descriptor) where T : IBeamableService =>
 		   GetModel<T>(descriptor.Name);
 
@@ -286,6 +292,7 @@ namespace Beamable.Editor.UI.Model
 	{
 		AllTypes,
 		Microservices,
-		Storages
+		Storages,
+		Archived
 	}
 }